+++ src/news/mod.rs
@@ -3,7 +3,9 @@ mod model;
 mod request;
 mod response;
 
+pub use crate::common::enums::Sort;
+
 pub use client::NewsClient;
-pub use model::NewsItem;
+pub use model::{NewsImage, NewsItem};
 pub use request::ListRequest;
 pub use response::ListResponse;